# openapi/openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HartDiseaseModelResponse:
      properties:
        prediction:
          title: Prediction class
          type: int
      title: HartDiseaseModelResponse
      type: object
    ValidationError:
      properties:
        loc:
          items:
            type: string
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: Currently provides an serving API for UCI Hart Diseases prediction
  title: UCI Hart Diseases prediction
  version: 0.1.0
openapi: 3.0.2
paths:
  /predict:
    get:
      description: This method returns the model predictions
      operationId: predict
      parameters:
      - in: query
        name: age
        required: true
        schema:
          exclusiveMinimum: 0
          title: Age of the patient
          type: integer
      - in: query
        name: sex
        required: true
        schema:
          title: Sex of the patient
          type: string
      - in: query
        name: dataset
        required: true
        schema:
          title: Geo location of the patient
          type: string
      - in: query
        name: cp
        required: true
        schema:
          title: Chest pain type of the patient
          type: string
      - in: query
        name: trestbps
        required: true
        schema:
          exclusiveMinimum: 0
          title: Resting blood pressure of the patient
          type: float
      - in: query
        name: chol
        required: true
        schema:
          exclusiveMinimum: 0
          title: Serum cholesterol in mg/dl of the patient
          type: float
      - in: query
        name: fbs
        required: true
        schema:
          title: If fasting blood sugar > 120 mg/dl
          type: boolean
      - in: query
        name: restecg
        required: true
        schema:
          title: Resting electrocardiographic results of the patient
          type: string
      - in: query
        name: thalch
        required: true
        schema:
          exclusiveMinimum: 0
          title: Maximum heart rate achieved of the patient
          type: float
      - in: query
        name: exang
        required: true
        schema:
          title: Exercise-induced angina of the patient
          type: boolean
      - in: query
        name: oldpeak
        required: true
        schema:
          exclusiveMinimum: 0
          title: ST depression induced by exercise relative to rest
          type: float
      - in: query
        name: slope
        required: true
        schema:
          title: The slope of the peak exercise ST segment of the patient
          type: string
      - in: query
        name: ca
        required: true
        schema:
          title: The number of major vessels of the patient
          type: int
          inclusiveMinimum: 0
          inclusiveMaximum: 3
      - in: query
        name: thal
        required: true
        schema:
          title: Thal of the patient
          type: string
      responses:
        '200':
          content:
            application/json:
              example:
                prediction: 3
              schema:
                $ref: '#/components/schemas/HartDiseaseModelResponse'
          description: Model predict for passed data
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get the the hart disease prediction 
      tags:
      - predict